Enhance maintainability
Normal java convention is to put all or most of the variable declarations at the top of the file instead of declaring in the middle of the code.

Changed all usages of float to double. There really isn't any reason to use float at all, but many reasons to use double. First reason is that double is the default type for java decimal numbers. So if you wanted to do float n = 0.0, that would not work without a cast. As you can see, switching to double got rid of many casts in the code which should simplify things. Also double has more accuracy than float. So yeah, no reason to be using float unless you're trying to save a miniscule amount of memory.